# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Jumyste application",
        "title": "Jumyste App API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/jobs/apply/{vacancy_id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Apply for a job by providing vacancy ID and user details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Job Applications"
                ],
                "summary": "Apply for a job",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vacancy ID",
                        "name": "vacancy_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.JobApplicationResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid vacancy ID",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to apply for job",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/jobs/{application_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a job application by application ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Job Applications"
                ],
                "summary": "Delete a job application",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Application ID",
                        "name": "application_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid application ID",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete job application",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/jobs/{application_id}/status/{status}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the status of a specific job application by application ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Job Applications"
                ],
                "summary": "Update the status of a job application",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Application ID",
                        "name": "application_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "New Status",
                        "name": "status",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid application ID or status",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to update application status",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/jobs/{vacancy_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all job applications for a specific vacancy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Job Applications"
                ],
                "summary": "Get job applications by vacancy ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vacancy ID",
                        "name": "vacancy_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/jumyste-app-backend_internal_dto.JobApplicationResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid vacancy ID",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve job applications",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "jumyste-app-backend_internal_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid input"
                }
            }
        },
        "jumyste-app-backend_internal_dto.JobApplicationResponse": {
            "type": "object",
            "properties": {
                "applied_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "vacancy_id": {
                    "type": "integer"
                }
            }
        },
...
        "jumyste-app-backend_internal_dto.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "User registered successfully"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}